upload

+++ Backend/JavaNC_BookWorld/src/main/java/net/javanc/JavaNC_BookWorld/repository/GenreRepository.java
@@ -3,4 +3,4 @@ package net.javanc.JavaNC_BookWorld.repository;
 import net.javanc.JavaNC_BookWorld.model.Genre;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GenreRepository extends JpaRepository<Genre, Integer> {}
+public interface GenreRepository extends JpaRepository<Genre, Long> {}
